moved the shopping cart code to ShoppingCart.js, and now App.js only renders a ShoppingCart element
